Share multicluster spoke E2E helpers
